backend/requirements.txt
Flask
Flask
Flask-SQLAlchemy
psycopg2-binary
pandas
scikit-learn
flask-cors
python-dotenv
gunicorn
flask-jwt-extended
bcrypt
requests
